+++ database/users/tokens.json
@@ -1,1 +1,1 @@
-[{"id":"7150a863-8bf9-424e-9e63-5e76ab821934","user":"cfc70b92-4f87-4311-82ae-70a7da32349a","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InVzZXI1QGdtYWlsLmNvbSIsImlkIjoiY2ZjNzBiOTItNGY4Ny00MzExLTgyYWUtNzBhN2RhMzIzNDlhIiwibG9naW4iOiJsb2dpbjUiLCJhZG1pbiI6ZmFsc2UsImlhdCI6MTY1MTY3MDY4MCwiZXhwIjoxNjUyNTM0NjgwfQ.-8GVgfdErx-YAF3OVszUGE5cC2h8Am16KTzHN9Sx4OM"}]
+[]
